move ut_solv, ut_reconstr to _solve, _reconstruct
